backend/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
python-jose[cryptography]
passlib[bcrypt]
python-multipart
transformers
torch
Pillow
